New test case added to validate the DB writing method.

+++ nqueens.py
@@ -7,13 +7,15 @@ import json
 @click.argument('nqueens')
 @click.option('--iterative', '-i', is_flag=True,\
 help = "Iterate over N and store the solutions in postgres")
-def main(nqueens,iterative):
+@click.option('--solutions', '-s', is_flag=True,\
+help = "Print all the solution for the N queen problem stored in postgres")
+def main(nqueens, iterative, solutions):
     """
     Application to determinate all the solutions for the N queens problem.
     This application reads the db parameter from the db_settings.json file.
     """
     n = int(nqueens)
-    if iterative:
+    if iterative or solutions:
         with open('db_settings.json') as json_file: #Read the configuration for the db.
             data = json.load(json_file)
             p = data["settings"]
@@ -23,16 +25,27 @@ def main(nqueens,iterative):
             port = p['port']
             database = p['database']
         DB = dbConnection(user,password,host,port,database)
-        for i in range(1,n+1): #Iterate over N to find the solutions and store the results
-            board = Board(i)
-            DB.dropTable("queens{}".format(i))
-            DB.createTable(i,"queens{}".format(i))
-            finds, solutions = board.findSolutions()
-            print(finds)
-            DB.wirteSolutions(solutions)
+        if iterative:
+            for i in range(1,n+1): #Iterate over N to find the solutions and store the results
+                board = Board(i)
+                DB.dropTable("queens{}".format(i))
+                DB.createTable(i,"queens{}".format(i))
+                finds, solutions = board.findSolutions()
+                print(finds)
+                DB.writeSolutions(solutions)
+        else:
+            solutions = DB.readSolutions("queens{}".format(n)) # Read the solutions stored in the DB
+            if solutions != None:
+                for i,queen_solution in enumerate(solutions):
+                    print(i+1,"->  ",queen_solution.queen_rows,"  ",queen_solution.queen_cols)
+            else:
+                print("Table for {0} queens problem not found. \
+                    \nRun the application with args. -i {0}  \
+                    \nto create this solution's table".format(n))
     else:
         board = Board(n) #Create an object of Board type to find the solution of the n queens problem
         finds, solutions = board.findSolutions()
+        solutions = DB.readSolutions("queens8".format(n))
         print(finds)
 
 if __name__ == "__main__":
